edit

+++ Form1.cs
@@ -167,25 +167,38 @@ namespace Diar
                 nazev = dialog.udalost;
                 if (priorita == 1)
                 {
-                    ctverec = "\u001b[31m█\u001b[0m"; //pro referenci: https://gist.github.com/dominikwilkowski/60eed2ea722183769d586c76f22098dd
+                    ctverec = "A";
                 }
                 else if (priorita == 2)
                 {
-                    ctverec = "\u001b[33m█\u001b[0m";
+                    ctverec = "B";
                 }
                 else if (priorita == 3)
                 {
-                    ctverec = "\u001b[32m█\u001b[0m";
+                    ctverec = "C";
                 }
                 conn.Open();
-                id = Convert.ToInt32(dataGridView1.SelectedRows[0].Cells["Id"].Value);
+                if (dataGridView1.SelectedRows.Count == 0)
+                {
+                    conn.Close();
+                    MessageBox.Show("Vyber celý řádek");
+                    return;
+                }
+                id = Convert.ToInt32(dataGridView1.SelectedRows[0].Cells["cislo"].Value);
                 sql = $"DELETE FROM udalosti where id = {id}";
                 cmd = new SQLiteCommand(sql, conn);
                 cmd.ExecuteNonQuery();
-                sql = $"INSERT INTO udalosti(id, nazev, interval, dulezitost, datum, alarm) VALUES({id}, {nazev}, {doba},{ctverec},{cas},{upozorneni} )";
+                sql = $"INSERT INTO udalosti(id, nazev, interval, dulezitost, datum, alarm) VALUES(@id, @nazev, @interval,@dulezitost,@datum,@alarm)";
                 cmd = new SQLiteCommand(sql, conn);
+                cmd.Parameters.AddWithValue("@id", id);
+                cmd.Parameters.AddWithValue("@nazev", nazev);
+                cmd.Parameters.AddWithValue("@interval", doba);
+                cmd.Parameters.AddWithValue("@dulezitost", ctverec);
+                cmd.Parameters.AddWithValue("@datum", cas);
+                cmd.Parameters.AddWithValue("@alarm", upozorneni);
                 cmd.ExecuteNonQuery();
                 operace = 2;
+                udalost udal = new udalost(cas, this, operace, upozorneni, nazev);
                 dataGridView1.DataSource = null;
                 sql = "SELECT * FROM udalosti ORDER BY id";
                 cmd = new SQLiteCommand(sql, conn);
